Reworked building ontology using only basic terms and properties
It acts more like a basement with basic information rather
than a full fledged package of classes and properties to
describe each building you can imagine.

The idea is, if you want to describe the architecture of
the building, e.g. which rooms are connected to each
other, which room is on which level, you may extend this
ontology with our own.